+++ docs/source/conf.py
@@ -4,23 +4,14 @@
 # list see the documentation:
 # https://www.sphinx-doc.org/en/master/usage/configuration.html
 
-# -- Path setup --------------------------------------------------------------
-
-# If extensions (or modules to document with autodoc) are in another directory,
-# add these directories to sys.path here. If the directory is relative to the
-# documentation root, use os.path.abspath to make it absolute, like shown here.
-#
-import os
-import sys
-sys.path.insert(0, os.path.abspath('.'))
-
+from datetime import date
 
 # -- Project information -----------------------------------------------------
 
-docstitle= 'sphinxcontrib-ghcontributors'
-project = 'sphinxcontrib-ghcontributors'
-copyright = '2021, David Garcia'
-author = 'David Garcia'
+docstitle = "sphinx-collapse"
+project = "sphinx-collapse"
+copyright = str(date.today().year) + ", David Garcia. All rights reserved."
+author = "David Garcia"
 
 
 # -- General configuration ---------------------------------------------------
@@ -28,17 +19,15 @@ author = 'David Garcia'
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
-extensions = [
-    'sphinxcontrib.ghcontributors'
-]
+extensions = ["sphinx_contributors"]
 
 # Add any paths that contain templates here, relative to this directory.
-templates_path = ['_templates']
+templates_path = ["_templates"]
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
-exclude_patterns = []
+exclude_patterns = ["build/*"]
 
 
 # -- Options for HTML output -------------------------------------------------
@@ -46,15 +35,15 @@ exclude_patterns = []
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-html_theme = 'furo'
+html_theme = "furo"
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
-# so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ['_static']
+# so a file named 'default.css' will overwrite the builtin 'default.css'.
+html_static_path = ["_static"]
 
 # These paths are either relative to html_static_path
 # or fully qualified paths (eg. https://...)
 html_css_files = [
-    'css/custom.css',
+    "css/custom.css",
 ]